+++ src/types.ts
@@ -14,9 +14,10 @@ export interface EnvCmdOptions extends GetEnvVarOptions {
   command: string
   commandArgs: string[]
   options?: {
+    expandEnvs?: boolean
     noOverride?: boolean
+    silent?: boolean
     useShell?: boolean
     verbose?: boolean
-    expandEnvs?: boolean
   }
 }