+++ src/parse-env-file.ts
@@ -18,7 +18,7 @@ export async function getEnvFileVars (envFilePath: string): Promise<{ [key: stri
   // Get the file extension
   const ext = path.extname(absolutePath).toLowerCase()
   let env = {}
-  if (REQUIRE_HOOK_EXTENSIONS.indexOf(ext) > -1) {
+  if (REQUIRE_HOOK_EXTENSIONS.includes(ext)) {
     const possiblePromise = require(absolutePath) /* eslint-disable-line */
     env = isPromise(possiblePromise) ? await possiblePromise : possiblePromise
   } else {